Added dataset_transformer.py
Created a class that could help deal with computationally expensive batch transforms where we want to just save the transformed data to file. This class was used to transform the tiny EFB dataset into a mirrored dataset where all the wav files are transformed into spectrograms. This changes the dataset from a set of wav files into a set of images -- (so that that computation does not need to happen each batch.